requirement.txt
PyYAML
flask
flask-restful
setproctitle
gunicorn
gevent
request
Flask-HTTPAuth
requests